feat: lesson-6 - react app structure
