+++ src/lesson-6/src/features/tasks/components/CreateTaskForm/index.tsx
@@ -2,11 +2,10 @@ import { ErrorMessage } from '@hookform/error-message';
 import { Controller, useForm } from 'react-hook-form';
 import { newTaskDefaultValueSchema } from './validationSchema';
 import cx from 'classnames';
-import { useNavigate } from 'react-router';
 import { taskApi } from '../../../../store/tasks';
-import type { FC } from 'react';
-import type { NewTaskBody } from '../../types';
-// import type { NewTaskBody } from '../../types';
+import type { NewTaskBody } from '../../../types';
+import { useNavigate } from 'react-router';
+import { AppRoutes } from '../../../../shared/types/router';
 
 const inputElementStyles = {
     label: 'block text-sm font-medium text-gray-700',
@@ -18,37 +17,37 @@ const inputElementStyles = {
         'w-full bg-green-600 hover:bg-green-700 disabled:bg-red-700 text-white font-semibold px-4 py-2 rounded-lg shadow-md',
 };
 
-export const CreateTaskForm: FC = () => {
+export const CreateTaskForm = () => {
     const [createTaskMutation] = taskApi.useCreateTaskMutation();
-    const navigate = useNavigate();
+    const navigation = useNavigate();
 
     const {
         register,
         handleSubmit,
         control,
         formState: { errors, isValid },
+        reset,
     } = useForm<NewTaskBody>({
         mode: 'all',
         defaultValues: newTaskDefaultValueSchema,
     });
 
     const onSubmit = async (formData: NewTaskBody) => {
-        await createTaskMutation(formData);
-        navigate({
-            pathname: '/task'
-        });
+        await createTaskMutation({ ...formData, createdAt: new Date() });
+        reset();
+        navigation(AppRoutes.TASK_LIST);
     };
 
     return (
-        <form className='w-full grid gap-2' onSubmit={handleSubmit(onSubmit)}>
+        <form className="w-full grid gap-2" onSubmit={handleSubmit(onSubmit)}>
             <div>
-                <label htmlFor='title' className={cx(inputElementStyles.label)}>
+                <label htmlFor="title" className={cx(inputElementStyles.label)}>
                     Title
                 </label>
                 <input
-                    type='text'
-                    id='title'
-                    placeholder='Title'
+                    type="text"
+                    id="title"
+                    placeholder="Title"
                     className={cx(inputElementStyles.label, inputElementStyles.default, {
                         [inputElementStyles.error]: errors.title,
                     })}
@@ -63,7 +62,7 @@ export const CreateTaskForm: FC = () => {
 
                 <ErrorMessage
                     errors={errors}
-                    name='title'
+                    name="title"
                     render={(data) => (
                         <span className={cx(inputElementStyles.errorText)}>
                             {data.message}
@@ -73,12 +72,12 @@ export const CreateTaskForm: FC = () => {
             </div>
 
             <div>
-                <label htmlFor='description' className={cx(inputElementStyles.label)}>
+                <label htmlFor="description" className={cx(inputElementStyles.label)}>
                     Description
                 </label>
                 <textarea
-                    id='description'
-                    placeholder='Description'
+                    id="description"
+                    placeholder="Description"
                     className={cx(inputElementStyles.label, inputElementStyles.default, {
                         [inputElementStyles.error]: errors.description,
                     })}
@@ -92,7 +91,7 @@ export const CreateTaskForm: FC = () => {
                 ></textarea>
                 <ErrorMessage
                     errors={errors}
-                    name='description'
+                    name="description"
                     render={(data) => (
                         <span className={cx(inputElementStyles.errorText)}>
                             {data.message}
@@ -102,23 +101,23 @@ export const CreateTaskForm: FC = () => {
             </div>
 
             <div>
-                <label htmlFor='status' className={cx(inputElementStyles.label)}>
+                <label htmlFor="status" className={cx(inputElementStyles.label)}>
                     Status
                 </label>
                 <select
-                    id='status'
+                    id="status"
                     className={cx(inputElementStyles.label, inputElementStyles.default, {
                         [inputElementStyles.error]: errors.status,
                     })}
                     {...register('status')}
                 >
-                    <option value='todo'>To do</option>
-                    <option value='in_progress'>In progress</option>
-                    <option value='done'>Done</option>
+                    <option value="todo">To do</option>
+                    <option value="in_progress">In progress</option>
+                    <option value="done">Done</option>
                 </select>
                 <ErrorMessage
                     errors={errors}
-                    name='status'
+                    name="status"
                     render={(data) => (
                         <span className={cx(inputElementStyles.errorText)}>
                             {data.message}
@@ -128,23 +127,23 @@ export const CreateTaskForm: FC = () => {
             </div>
 
             <div>
-                <label htmlFor='priority' className={cx(inputElementStyles.label)}>
+                <label htmlFor="priority" className={cx(inputElementStyles.label)}>
                     Priority
                 </label>
                 <select
-                    id='priority'
+                    id="priority"
                     className={cx(inputElementStyles.label, inputElementStyles.default, {
                         [inputElementStyles.error]: errors.priority,
                     })}
                     {...register('priority')}
                 >
-                    <option value='low'>Low</option>
-                    <option value='normal'>Normal</option>
-                    <option value='high'>High</option>
+                    <option value="low">Low</option>
+                    <option value="normal">Normal</option>
+                    <option value="high">High</option>
                 </select>
                 <ErrorMessage
                     errors={errors}
-                    name='priority'
+                    name="priority"
                     render={(data) => (
                         <span className={cx(inputElementStyles.errorText)}>
                             {data.message}
@@ -154,12 +153,12 @@ export const CreateTaskForm: FC = () => {
             </div>
 
             <div>
-                <label htmlFor='deadline' className={cx(inputElementStyles.label)}>
+                <label htmlFor="deadline" className={cx(inputElementStyles.label)}>
                     Deadline
                 </label>
                 <Controller
                     control={control}
-                    name='deadline'
+                    name="deadline"
                     rules={{
                         required: 'Deadline is required',
                         validate: (value) => {
@@ -174,8 +173,8 @@ export const CreateTaskForm: FC = () => {
                     render={({ field: { onChange } }) => (
                         <>
                             <input
-                                type='date'
-                                id='deadline'
+                                type="date"
+                                id="deadline"
                                 className={cx(
                                     inputElementStyles.label,
                                     inputElementStyles.default,
@@ -190,7 +189,7 @@ export const CreateTaskForm: FC = () => {
                             />
                             <ErrorMessage
                                 errors={errors}
-                                name='deadline'
+                                name="deadline"
                                 render={(data) => (
                                     <span className={cx(inputElementStyles.errorText)}>
                                         {data.message}
@@ -203,7 +202,7 @@ export const CreateTaskForm: FC = () => {
             </div>
 
             <button
-                type='submit'
+                type="submit"
                 className={cx(inputElementStyles.submitButton)}
                 disabled={!isValid}
             >
