+++ src/lesson-6/src/features/tasks/components/CreateTaskForm/validationSchema.ts
@@ -1,10 +1,9 @@
-import type { NewTaskBody } from '../../types';
+import type { NewTaskBody } from '../../../types';
 
 export const newTaskDefaultValueSchema: NewTaskBody = {
   title: '',
   description: '',
   status: 'todo',
-  deadline: new Date().toISOString(),
+  deadline: new Date(),
   priority: 'low',
-  createdAt: new Date().toISOString(),
 };
