+++ src/lesson-6/src/features/types/models/bug.model.ts
@@ -1,0 +1,22 @@
+import { TaskType, type TaskBody, Priority } from '..';
+import { Task } from './task.model';
+
+export class Bug extends Task {
+  type = TaskType.bug;
+
+  constructor(
+    data: TaskBody,
+    public parentId: string | null = null,
+    public priority: Priority = Priority.low,
+  ) {
+    super(
+      data.title,
+      data.description,
+      data.status,
+      data.deadline,
+      data.createdAt,
+      priority,
+    );
+    this.id = data.id ?? crypto.randomUUID();
+  }
+}
