+++ src/lesson-6/src/features/types/models/epic.model.ts
@@ -1,0 +1,29 @@
+import { TaskType, type TaskBody } from '..';
+import { Task } from './task.model';
+
+export class Epic extends Task {
+  type = TaskType.epic;
+
+  constructor(
+    data: TaskBody,
+    public children: string[] = [],
+  ) {
+    super(
+      data.title,
+      data.description,
+      data.status,
+      data.deadline,
+      data.createdAt,
+      data.priority,
+    );
+    this.id = data.id ?? crypto.randomUUID();
+  }
+
+  addChild(childId: string) {
+    this.children.push(childId);
+  }
+
+  getChildren() {
+    return this.children;
+  }
+}
